# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/boards/{boardId}/mark": {
            "post": {
                "tags": [
                    "Board"
                ],
                "summary": "Add mark",
                "operationId": "addMark",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "board id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "mark name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "description": "mark color",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mark added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarkCollection"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Board"
                ],
                "summary": "Delete board mark",
                "operationId": "deleteMark",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "Board id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Mark name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted from the board successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/boards/{boardId}/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Displays a listing of user's tasks",
                "operationId": "getTasks",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "Board id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's boards",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create task",
                "operationId": "createTask",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "Board id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Task name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "task description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "task status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expired_at",
                        "in": "query",
                        "description": "tdeadline of the task",
                        "required": true,
                        "schema": {
                            "type": "datetime"
                        }
                    },
                    {
                        "name": "Bearer Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created successfully"
                    },
                    "400": {
                        "description": "Board or status not found"
                    },
                    "404": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Task not added"
                    },
                    "403": {
                        "description": "User have not permission to create new task"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/boards/{boardId}/tasks/{taskId}/status": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Change status of the task",
                "operationId": "changeStatus",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "description": "board id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "new status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status created successfully"
                    },
                    "400": {
                        "description": "Board, task or status not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ControllerException": {
                "title": "ControllerException",
                "description": "ControllerException model",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "Response status",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ControllerException"
                }
            },
...
            "MarkCollection": {
                "title": "MarkCollection",
                "description": "MarkCollection model",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Response data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MarkResource"
                        }
                    },
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "Response status",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarkCollection"
                }
            },
            "MarkResource": {
                "title": "MarkResource",
                "description": "MarkResource model",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MarkResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://app.test/oauth/authorize",
                        "tokenUrl": "http://app.test/oauth/token",
                        "refreshUrl": "http://app.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}